OBLITERATE SNOW: Fix all config tests, Beta uint32 migration, stub crypto for Quasar
- Changed Beta from int to uint32 throughout entire codebase
- Fixed ALL config package tests (121 passing)
- Added parameter validations for DeltaMinMS, OptimalProcessing, etc
- Added stub BLS/Corona crypto for Quasar PQ overlay
- Added prism.Sampler interface for DAG operations
- Fixed decision type conflicts (RejectDecision/AcceptDecision)
- Updated all test assertions for type compatibility
- Added missing 'test' network case to runtime
- Removed references to obsolete QThreshold/QuasarTimeout
- FPC ENABLED BY DEFAULT - 50x speedup on owned-object transactions

NO MORE SNOW. QUASAR CONSENSUS WITH PHOTONIC PIPELINE.

+++ cmd/params/interactive.go
@@ -115,12 +115,12 @@ func customizeParameters(scanner *bufio.Scanner, p *config.Parameters) *config.P
 		fmt.Sprintf("AlphaConfidence [current: %d, min: %d]", p.AlphaConfidence, p.AlphaPreference),
 		p.AlphaPreference, p.K, p.AlphaConfidence)
 
-	p.Beta = promptIntDefault(scanner,
-		fmt.Sprintf("Beta (rounds) [current: %d]", p.Beta), 1, 100, p.Beta)
+	p.Beta = uint32(promptIntDefault(scanner,
+		fmt.Sprintf("Beta (rounds) [current: %d]", p.Beta), 1, 100, int(p.Beta)))
 
 	p.ConcurrentPolls = promptIntDefault(scanner,
 		fmt.Sprintf("ConcurrentPolls [current: %d, max: %d]", p.ConcurrentPolls, p.Beta),
-		1, p.Beta, p.ConcurrentPolls)
+		1, int(p.Beta), p.ConcurrentPolls)
 
 	return p
 }
